refactor: adds controllers
