+++ src/modules/users/infra/http/routes/users.routes.ts
@@ -1,30 +1,10 @@
 import { Router } from 'express';
-import { container } from 'tsyringe';
 
-import CreateUsersService from '@modules/users/services/CreateUserService';
+import UserController from '@modules/users/infra/http/controllers/UserController';
 
 const router = Router();
+const userController = new UserController();
 
-interface IPostBody {
-  name: string;
-  nick_name: string;
-  email: string;
-  password: string;
-}
-
-router.post('', async (req, res) => {
-  const { name, nick_name, email, password } = req.body as IPostBody;
-
-  const createUser = container.resolve(CreateUsersService);
-
-  const user = await createUser.execute({
-    name,
-    email,
-    nick_name,
-    password,
-  });
-
-  return res.json(user);
-});
+router.post('', userController.create);
 
 export default router;
